# chore: add skeleton for the stack-monitoring tests