--- a/.ci/Jenkinsfile
+++ b/.ci/Jenkinsfile
@@ -98,7 +98,7 @@ pipeline {
               }
               post {
                 always {
-                  junit(allowEmptyResults: true, keepLongStdio: true, testResults: "${BASE_DIR}/outputs/TEST-unit.xml")
+                  junit(allowEmptyResults: true, keepLongStdio: true, testResults: "${BASE_DIR}/outputs/TEST-unit*.xml")
                 }
               }
             }