--- a/.ci/scripts/build-test.sh
+++ b/.ci/scripts/build-test.sh
@@ -19,4 +19,7 @@ export OUT_FILE="outputs/test-report.out"
 make -C cli install test | tee ${OUT_FILE}
 
 go get -v -u github.com/jstemmer/go-junit-report
-go-junit-report > outputs/TEST-unit.xml < ${OUT_FILE}
+go-junit-report > outputs/TEST-unit-cli.xml < ${OUT_FILE}
+
+make -C e2e test | tee ${OUT_FILE}
+go-junit-report > outputs/TEST-unit-e2e.xml < ${OUT_FILE}