--- a/e2e/configurations/parity-testing/pipelines/parity.conf
+++ b/e2e/configurations/parity-testing/pipelines/parity.conf
@@ -1,0 +1,15 @@
+input {
+  file {
+    path => "/var/log/syslog"
+    start_position => "beginning"
+  }
+}
+output {
+  elasticsearch {
+    hosts => ["${monitoringEsHost:-elasticsearch}:${monitoringEsPort:-9200}"]
+    user => "elastic"
+    password => "p4ssw0rd"
+    ssl => "false"
+  }
+  stdout { codec => rubydebug }
+}