--- a/e2e/go.mod
+++ b/e2e/go.mod
@@ -12,6 +12,7 @@ require (
 	github.com/onsi/ginkgo v1.10.1 // indirect
 	github.com/onsi/gomega v1.7.0 // indirect
 	github.com/sirupsen/logrus v1.4.2
+	github.com/stretchr/testify v1.5.1
 )
 
 replace github.com/elastic/e2e-testing/cli v1.0.0-rc1 => ../cli