--- a/e2e/stack_monitoring_checks_test.go
+++ b/e2e/stack_monitoring_checks_test.go
@@ -1,3 +1,5 @@
+// +build unit
+
 package e2e
 
 import (