… .vscode/launch.json
… cli/config/compose/services/elasticsearch/docker-compose.yml
… cli/config/compose/services/kibana/docker-compose.yml
… cli/config/compose/services/logstash/docker-compose.yml
… e2e/README.md
… e2e/configurations/metricbeat/monitoring-beat.yml
… e2e/configurations/metricbeat/monitoring-elasticsearch.yml
… e2e/configurations/metricbeat/monitoring-filebeat.yml
… e2e/configurations/metricbeat/monitoring-kibana.yml
… e2e/configurations/metricbeat/monitoring-logstash.yml
… e2e/go.sum
… e2e/runner_test.go
… e2e/stack_monitoring_test.go
… e2e/testresources/elasticsearch-legacy-monitoring.json
… e2e/testresources/elasticsearch-metricbeat-monitoring.json
… e2e/testresources/filebeat-legacy-monitoring.json
… e2e/testresources/filebeat-metricbeat-monitoring.json
… e2e/testresources/kibana-legacy-monitoring.json
… e2e/testresources/kibana-metricbeat-monitoring.json
… e2e/testresources/logstash-legacy-monitoring.json
… e2e/testresources/logstash-metricbeat-monitoring.json