## What is this PR doing?
It ports the existing Python parity tests for stack monitoring to Go

## Why is it important?
We want to add support for Stack Monitoring in this test framework.

## How to test these changes?
You could run: 

```shell
$ cd e2e
$ OP_LOG_LEVEL=INFO go test -v --godog.format junit parity-tests
```

or using VSCode, supporting debug capabilities:
- Add breakpoints in the functions you're interested
- Open runner_test.go file as the current selected file in the editor
- Open Debug panel
- Select "Parity Tests" in the Run selector

## Related issues
- Closes #94

## Logs

<details><summary>Expand to view the logs</summary>
<p>
<pre><code>
OP_LOG_LEVEL=DEBUG go test -v --godog.format pretty parity-tests
DEBU[0000] Validating required tools: [docker docker-compose] 
DEBU[0000] Binary is present                             binary=docker path=/usr/local/bin/docker
DEBU[0000] Binary is present                             binary=docker-compose path=/usr/local/bin/docker-compose
DEBU[0000] 'op' workdirs created.                        servicesPath=/Users/mdelapenya/.op/compose/services stacksPath=/Users/mdelapenya/.op/compose/stacks
DEBU[0000] Boxed file                                    path=services/apm-server/docker-compose.yml service=apm-server
DEBU[0000] Boxed file                                    path=services/elasticsearch/docker-compose.yml service=elasticsearch
DEBU[0000] Boxed file                                    path=services/kibana/docker-compose.yml service=kibana
DEBU[0000] Boxed file                                    path=services/metricbeat/docker-compose.yml service=metricbeat
DEBU[0000] Boxed file                                    path=services/opbeans-go/docker-compose.yml service=opbeans-go
DEBU[0000] Boxed file                                    path=services/opbeans-java/docker-compose.yml service=opbeans-java
DEBU[0000] Boxed file                                    path=services/vsphere/docker-compose.yml service=vsphere
DEBU[0000] Boxed file                                    path=stacks/metricbeat/docker-compose.yml service=metricbeat
DEBU[0000] Workspace file                                path=/Users/mdelapenya/.op/compose/services/apache/docker-compose.yml service=apache
DEBU[0000] Workspace file                                path=/Users/mdelapenya/.op/compose/services/ceph/docker-compose.yml service=ceph
DEBU[0000] Workspace file                                path=/Users/mdelapenya/.op/compose/services/dropwizard/docker-compose.yml service=dropwizard
DEBU[0000] Workspace file                                path=/Users/mdelapenya/.op/compose/services/envoyproxy/docker-compose.yml service=envoyproxy
DEBU[0000] Workspace file                                path=/Users/mdelapenya/.op/compose/services/etcd/docker-compose.yml service=etcd
DEBU[0000] Workspace file                                path=/Users/mdelapenya/.op/compose/services/haproxy/docker-compose.yml service=haproxy
DEBU[0000] Workspace file                                path=/Users/mdelapenya/.op/compose/services/kafka/docker-compose.yml service=kafka
DEBU[0000] Workspace file                                path=/Users/mdelapenya/.op/compose/services/metricbeat/docker-compose.yml service=metricbeat
DEBU[0000] Workspace file                                path=/Users/mdelapenya/.op/compose/services/mysql/docker-compose.yml service=mysql
DEBU[0000] Workspace file                                path=/Users/mdelapenya/.op/compose/services/nats/docker-compose.yml service=nats
DEBU[0000] Workspace file                                path=/Users/mdelapenya/.op/compose/services/redis/docker-compose.yml service=redis
DEBU[0000] Workspace file                                path=/Users/mdelapenya/.op/compose/services/uwsgi/docker-compose.yml service=uwsgi
DEBU[0000] Workspace file                                path=/Users/mdelapenya/.op/compose/services/vsphere/docker-compose.yml service=vsphere
DEBU[0000] Workspace file                                path=/Users/mdelapenya/.op/compose/stacks/metricbeat/docker-compose.yml service=metricbeat
INFO[0000] Feature Context found                         modules="[stack-monitoring]" paths="[features/stack-monitoring/parity-tests.feature]"
DEBU[0000] Before StackMonitoring Suite...              
DEBU[0000] Installing elasticsearch monitoring instance 
Feature: Parity Tests
DEBU[0000] Before StackMonitoring Scenario...           
DEBU[0000] Installing elasticsearch                     
DEBU[0000] Enabling legacy collection, sending metrics to the monitoring instance 
DEBU[0000] Running elasticsearch for X seconds (default: 30) to collect monitoring data internally and index it into the Monitoring index for elasticsearch 
DEBU[0000] Stopping elasticsearch                       
DEBU[0000] Downloading sample documents from elasticsearch's monitoring index to a test directory 
DEBU[0000] Disable legacy                               

  Scenario Outline: The <product> documents indexed by the legacy collection method are identical in structure to those indexed by Metricbeat collection # features/stack-monitoring/parity-tests.feature:3
    Given "<product>" sends metrics to Elasticsearch using the "legacy" collection monitoring method                                                     # stack_monitoring_test.go:15 -> *StackMonitoringTestSuite
    When "<product>" sends metrics to Elasticsearch using the "metricbeat" collection monitoring method                                                  # stack_monitoring_test.go:15 -> *StackMonitoringTestSuite
    Then the structure of the documents for the "legacy" and "metricbeat" collection are identical                                                       # stack_monitoring_test.go:50 -> *StackMonitoringTestSuite

    Examples:
      | product       |
      | elasticsearch |
DEBU[0000] After StackMonitoring Scenario...            
DEBU[0000] Before StackMonitoring Scenario...           
DEBU[0000] Installing kibana                            
DEBU[0000] Enabling legacy collection, sending metrics to the monitoring instance 
DEBU[0000] Running kibana for X seconds (default: 30) to collect monitoring data internally and index it into the Monitoring index for kibana 
DEBU[0000] Stopping kibana                              
DEBU[0000] Downloading sample documents from kibana's monitoring index to a test directory 
DEBU[0000] Disable legacy                               
      | kibana        |
DEBU[0000] After StackMonitoring Scenario...            
DEBU[0000] Before StackMonitoring Scenario...           
DEBU[0000] Installing beats                             
DEBU[0000] Enabling legacy collection, sending metrics to the monitoring instance 
DEBU[0000] Running beats for X seconds (default: 30) to collect monitoring data internally and index it into the Monitoring index for beats 
DEBU[0000] Stopping beats                               
DEBU[0000] Downloading sample documents from beats's monitoring index to a test directory 
DEBU[0000] Disable legacy                               
      | beats         |
DEBU[0000] After StackMonitoring Scenario...            
DEBU[0000] Before StackMonitoring Scenario...           
DEBU[0000] Installing logstash                          
DEBU[0000] Enabling legacy collection, sending metrics to the monitoring instance 
DEBU[0000] Running logstash for X seconds (default: 30) to collect monitoring data internally and index it into the Monitoring index for logstash 
DEBU[0000] Stopping logstash                            
DEBU[0000] Downloading sample documents from logstash's monitoring index to a test directory 
DEBU[0000] Disable legacy                               
      | logstash      |
DEBU[0000] After StackMonitoring Scenario...            
DEBU[0000] After StackMonitoring Suite...               
DEBU[0000] Destroying elasticsearch monitoring instance, including attached services 

4 scenarios (4 pending)
12 steps (4 pending, 8 skipped)
1.28132ms
testing: warning: no tests to run
PASS
ok      github.com/elastic/metricbeat-tests-poc/e2e     0.790s
</code></pre>
</p>
</details>