@ycombinator 2020-05-11
Hi @mdelapenya (and maybe also @jsoriano @andresrc) , quick question about this PR. I saw the recent email about closing the modules testing working group. Just want to double check that:
a. we want to continue the investment in the E2E testing prototype via this PR, and
b. if so, we are **also** committing to running the tests being implemented by this PR in CI, so we can deprecate the existing stack monitoring parity tests currently running in CI (e.g. https://internal-ci.elastic.co/job/elastic+estf-monitoring-snapshots+master/).

In other words, just re-implementing the _existing_ parity tests in the E2E testing prototype will be wasteful, IMO, if we don't also complete the picture by running them in CI just like the existing ones are. I just want to double check that this is the plan here before spending more time on this path.