@mdelapenya 2020-05-11
> In other words, just re-implementing the _existing_ parity tests in the E2E testing prototype will be wasteful, IMO, if we don't also complete the picture by running them in CI just like the existing parity tests. I just want to double check that this is the plan here before spending more time on this path.

Yes, in terms of stack-monitoring, once we have something that we like, the idea is to put that code in the CI and run it with the same frequency as it runs now. As you know, this repo is only run whenever there are changes on it, but we can integrate it with any PR on beats, or logstash or elasticsearch if we want.

This PR is an example, https://github.com/elastic/beats/pull/13073 (very old in time, but not in behavior): we can monitor anything we need and run this project in consequence