@mdelapenya 2020-05-12
Hey @ycombinator, thanks for you feedback, I've addressed it in the last commit.

What I'm concerned about is that for elasticsearch, kibana and logstash, the monitoring indices are not present:

```log
DEBU[0072] Elasticsearch query                           index=.monitoring-elasticsearch-7-2020.05.12 query="map[collapse:map[field:type] sort:map[timestamp:asc]]"
ERRO[0072] Could not send query to Elasticsearch in the specified time (15 seconds)  error="Could not send query to Elasticsearch in the specified time (15 seconds)" index=.monitoring-elasticsearch-7-2020.05.12 query="map[collapse:map[field:type] sort:map[timestamp:asc]]" retryAttempts=5 retryTimeout=3
DEBU[0072] Deleting monitoring index .monitoring-elasticsearch-7-2020.05.12 
DEBU[0072] Index deleted using Elasticsearch Go client   indexName=.monitoring-elasticsearch-7-2020.05.12 status="404 Not Found"

...

DEBU[0123] Elasticsearch query                           index=.monitoring-kibana-7-2020.05.12 query="map[collapse:map[field:type] sort:map[timestamp:asc]]"
ERRO[0123] Could not send query to Elasticsearch in the specified time (15 seconds)  error="Could not send query to Elasticsearch in the specified time (15 seconds)" index=.monitoring-kibana-7-2020.05.12 query="map[collapse:map[field:type] sort:map[timestamp:asc]]" retryAttempts=5 retryTimeout=3
DEBU[0123] Deleting monitoring index .monitoring-kibana-7-2020.05.12 
DEBU[0123] Index deleted using Elasticsearch Go client   indexName=.monitoring-kibana-7-2020.05.12 status="404 Not Found"

...

DEBU[0169] Elasticsearch query                           index=.monitoring-logstash-7-2020.05.12 query="map[collapse:map[field:type] sort:map[timestamp:asc]]"
ERRO[0169] Could not send query to Elasticsearch in the specified time (15 seconds)  error="Could not send query to Elasticsearch in the specified time (15 seconds)" index=.monitoring-logstash-7-2020.05.12 query="map[collapse:map[field:type] sort:map[timestamp:asc]]" retryAttempts=5 retryTimeout=3
DEBU[0169] Deleting monitoring index .monitoring-logstash-7-2020.05.12 
DEBU[0169] Index deleted using Elasticsearch Go client   indexName=.monitoring-logstash-7-2020.05.12 status="404 Not Found"
```