@ycombinator 2020-05-21
Hi @mdelapenya, I pulled down this PR and ran it. I focussed just on the Elasticsearch parity tests, not the other products. I did this because these tests are failing for me (and I believe for you as well) with this error:

```
        Found more legacy-indexed document types than metricbeat-indexed document types. Document types indexed by legacy collection but not by Metricbeat collection: [index_stats indices_stats shards]
``` 

As you know, the setup is as follows:
- The monitoring ES cluster runs on `localhost:9200`.
- The production ES cluster runs on `localhost:9201`.

At the end of the legacy collection stage, we expect monitoring documents to end up in `.monitoring-es-*` indices on the monitoring cluster. I was examining these documents and I noticed something wrong. The documents have a `cluster_uuid` field which **should** have the cluster UUID of the **production** ES cluster. Instead they contained the cluster UUID of the **monitoring** ES cluster. 

You can check this as follows:
1. Grab the cluster UUID of the monitoring ES cluster: 
   ```
   curl -s 'http://localhost:9200/?filter_path=cluster_uuid' | grep cluster_uuid
   ```
2. Grab the cluster UUID of the production ES cluster:
   ```
   curl -s 'http://localhost:9201/?filter_path=cluster_uuid' | grep cluster_uuid
   ```
3. Check the value of the `cluster_uuid` field in various documents in `.monitoring-es-*` indices on the monitoring ES cluster:
   ```
   curl -s 'http://localhost:9200/.monitoring-es-7-*/_search?filter_path=**._source.cluster_uuid&size=1000&pretty' | grep cluster_uuid | sort -u
   ```

To me, this indicates that the monitoring ES cluster is currently monitoring itself via legacy collection, instead of the production cluster monitoring itself via legacy collection and shipping it's data to the monitoring ES cluster.  So something is going wrong in how we are setting up the monitoring ES cluster and the production ES cluster both for legacy collection.