@ycombinator 2020-05-21
@mdelapenya Thinking about this some more, I think we are complicating the test implementation more than we need to. The issue that led to this PR is https://github.com/elastic/e2e-testing/issues/94. As I expressed in that issue, I really just want us to start by porting over **one** product's parity tests. Porting over, to me, also means keeping more or less the same steps as being done by the current parity tests in the new one here in the E2E testing framework. We can always make things more sophisticated later.

For instance, here are two examples of additional complexity we are taking on in this PR that I think we can avoid for now:

1. The current parity tests only spin up a single Elasticsearch cluster. It acts as the production and monitoring cluster both. The new ones in this PR are trying to spin up two Elasticsearch clusters, one to act as the production one and one to act as the monitoring one. Let's not tackle this complexity just yet. Let's simplify for now and just do what the existing parity tests are doing.

2. Another thing is that the current parity tests setup monitoring settings in Elasticsearch via API calls. In contrast, the new ones in this PR are trying to do this statically via environment variables passed to the docker containers. In theory this difference should not matter but I'd rather not introduce differences while we're porting over the tests. We can make these changes later if we believe they are necessary.

In short, I would really like the **initial version** of the new parity tests in the E2E framework to **just mimic exactly** the [steps](https://github.com/elastic/elastic-stack-testing/blob/master/playbooks/monitoring/elasticsearch/docs_parity.yml#L44) being done by the current parity tests. And let's just focus on **one product: Elasticsearch** for now as well.  Doing both these things will make debugging a whole lot easier IMO. Once we have a working PR, we can then move on making things more sophisticated in the future (like separating out the monitoring cluster from the production cluster).