@mdelapenya 2020-05-27
> However, performance was only one of the goals stated in #94. The other one was maintainability. Given that the @elastic/stack-monitoring-ui team is currently maintaining these parity tests, I will defer to them on whether they think porting them over to the E2E framework (as illustrated in this PR) would be more maintainable for them in the long run over the current implementation.

About performance, I'd run benchmarks for both implementations to verify it in a controlled environment to avoid getting simplistic results.

As an example, in my local machine (Macbook 13'', 16GB RAM, 2,7 GHz Intel Core i7, SSD) it takes:
```
ok      github.com/elastic/e2e-testing/e2e      195.666s
OP_LOG_LEVEL=INFO go test -v --godog.format junit parity-tests  6.06s user 2.75s system 4% cpu 3:18.55 total
```
which is faster than the above numbers (curious about the CPU usage: 4% vs 26%).

As a side note, the tests for the 4 products (elasticsearch, kibana, logstash and filebeat) complete  fairly fast:
```
FAIL    github.com/elastic/e2e-testing/e2e      576.602s
OP_LOG_LEVEL=INFO go test -v --godog.format junit parity-tests  14.35s user 4.28s system 3% cpu 9:38.84 total
```

Talking about **maintainability**, I'm going to list the things I (and the software industry via ISO's) considered important when working in both projects, creating this framework and reading the current parity tests. Hope I don't get anybody bored here 😄 

### Testability
- This framework:
   - Unit testing possible using programming language primitives (go test)
   - Language compiler supporting type checking, and code style
   - JSON processing relying on a third-party library, plus a few helper methods
   - Test Life cycle hooks (befores, afters) provided by the test libraries, being possible to create and destroy test resources using them.
- Current parity tests:
   - Linting of YAML difficult (indentations could be hell)
   - Unit tests for Ansible possible (i.e. [molecule](https://molecule.readthedocs.io/en/latest/)), although not present
   - Native JSON support in Python is great

### Analyzability
- This framework:
   - Golang as sole language for control flow
   - Golang code is organised in one single package, with files per functionality
   - Configuration variables defined in code, being possible to override them in the shell (i.e. OP_ variables).
- Current parity tests:
   - Ansible for control flow
   - YAML for Ansible role and playbook definition as main language
   - Ansible layout includes roles, playbooks, dependant roles
   - Configuration variables defined at Ansible level in multiple places: group_vars, host_vars, each role, playbook, shell (i.e. AIT_ variables).

### Functionality
- This framework:
   - Suitability:
     - It uses a high level, general purpose programming language (Golang), which seems more appropriate to create assertions
- Current parity tests:
   - Suitability:
     - It uses a declarative language (YAML) and Ansible, which uses YAML to define the state of the software to be installed in a target machine. It's possible to create assertions in YAML files.

### Usability
- This framework:
   - Learnability:
     - Golang is relatively simple to learn and understand, so any developer coming from a backend language would find most of language capabilities present in the language.
- Current parity tests:
   - Learnability:
     - YAML as declarative syntax is easy to learn, although difficult to maintain

### Installability
- This framework needs:
   - Golang 1.13
   - Docker
   - Docker Compose
- Current parity tests needs:
   - Python 3.6 (it does not run locally on my Mac with Python 3.7)
   - Virtualenv
   - Ansible
   - Vagrant
   - VirtualBox
   - Bash (to run shell scripts)